feat: payload header and memory arena class header
